requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
pydantic==2.5.3
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
apscheduler==3.10.4
pyyaml==6.0.1